+++ oss2/resumable.py
@@ -17,15 +17,14 @@ from . import defaults
 from .models import PartInfo
 from .compat import json, stringify, to_unicode
 from .task_queue import TaskQueue
+from .defaults import get_logger
 
-
-import logging
 import functools
 import threading
-import shutil
 import random
 import string
 
+import shutil
 
 _MAX_PART_COUNT = 10000
 _MIN_PART_SIZE = 100 * 1024
@@ -188,7 +187,7 @@ class _ResumableOperation(object):
 
         self.__store = store
         self.__record_key = self.__store.make_store_key(bucket.bucket_name, key, self._abspath)
-        logging.info('key is {0}'.format(self.__record_key))
+        get_logger().info('key is {0}'.format(self.__record_key))
 
         # protect self.__progress_callback
         self.__plock = threading.Lock()
@@ -286,7 +285,7 @@ class _ResumableDownloader(_ResumableOperation):
             headers = {'If-Match': self.objectInfo.etag,
                        'If-Unmodified-Since': utils.http_date(self.objectInfo.mtime)}
             result = self.bucket.get_object(self.key, byte_range=(part.start, part.end - 1), headers=headers)
-            shutil.copyfileobj(result, f)
+            utils.copyfileobj_and_verify(result, f, part.end - part.start, request_id=result.request_id)
 
         self.__finish_part(part)
 
@@ -332,20 +331,20 @@ class _ResumableDownloader(_ResumableOperation):
         try:
             for key in ('etag', 'tmp_suffix', 'abspath', 'bucket', 'key'):
                 if not isinstance(record[key], str):
-                    logging.info('{0} is not a string: {1}, but {2}'.format(key, record[key], record[key].__class__))
+                    get_logger().info('{0} is not a string: {1}, but {2}'.format(key, record[key], record[key].__class__))
                     return False
 
             for key in ('part_size', 'size', 'mtime'):
                 if not isinstance(record[key], int):
-                    logging.info('{0} is not an integer: {1}, but {2}'.format(key, record[key], record[key].__class__))
+                    get_logger().info('{0} is not an integer: {1}, but {2}'.format(key, record[key], record[key].__class__))
                     return False
 
             for key in ('parts'):
                 if not isinstance(record['parts'], list):
-                    logging.info('{0} is not a list: {1}, but {2}'.format(key, record[key], record[key].__class__))
+                    get_logger().info('{0} is not a list: {1}, but {2}'.format(key, record[key], record[key].__class__))
                     return False
         except KeyError as e:
-            logging.info('Key not found: {0}'.format(e.args))
+            get_logger().info('Key not found: {0}'.format(e.args))
             return False
 
         return True
@@ -356,7 +355,7 @@ class _ResumableDownloader(_ResumableOperation):
             record['etag'] != self.objectInfo.etag)
 
     def __finish_part(self, part):
-        logging.debug('finishing part: part_number={0}, start={1}, end={2}'.format(part.part_number, part.start, part.end))
+        get_logger().debug('finishing part: part_number={0}, start={1}, end={2}'.format(part.part_number, part.start, part.end))
 
         with self.__lock:
             self.__finished_parts.append(part)
@@ -464,12 +463,12 @@ class _ResumableUploader(_ResumableOperation):
             record = None
 
         if record and self.__file_changed(record):
-            logging.debug('{0} was changed, clear the record.'.format(self.filename))
+            get_logger().debug('{0} was changed, clear the record.'.format(self.filename))
             self._del_record()
             record = None
 
         if record and not self.__upload_exists(record['upload_id']):
-            logging.debug('{0} upload not exist, clear the record.'.format(record['upload_id']))
+            get_logger().debug('{0} upload not exist, clear the record.'.format(record['upload_id']))
             self._del_record()
             record = None
 
@@ -480,7 +479,7 @@ class _ResumableUploader(_ResumableOperation):
                       'abspath': self._abspath, 'bucket': self.bucket.bucket_name, 'key': self.key,
                       'part_size': part_size}
 
-            logging.debug('put new record upload_id={0} part_size={1}'.format(upload_id, part_size))
+            get_logger().debug('put new record upload_id={0} part_size={1}'.format(upload_id, part_size))
             self._put_record(record)
 
         self.__record = record
@@ -546,7 +545,7 @@ class _ResumableStoreBase(object):
     def get(self, key):
         pathname = self.__path(key)
 
-        logging.debug('get key={0}, pathname={1}'.format(key, pathname))
+        get_logger().debug('get key={0}, pathname={1}'.format(key, pathname))
 
         if not os.path.exists(pathname):
             return None
@@ -569,13 +568,13 @@ class _ResumableStoreBase(object):
         with open(to_unicode(pathname), 'w') as f:
             json.dump(value, f)
 
-        logging.debug('put key={0}, pathname={1}'.format(key, pathname))
+        get_logger().debug('put key={0}, pathname={1}'.format(key, pathname))
 
     def delete(self, key):
         pathname = self.__path(key)
         os.remove(pathname)
 
-        logging.debug('del key={0}, pathname={1}'.format(key, pathname))
+        get_logger().debug('del key={0}, pathname={1}'.format(key, pathname))
 
     def __path(self, key):
         return os.path.join(self.dir, key)
@@ -656,23 +655,23 @@ def _is_record_sane(record):
     try:
         for key in ('upload_id', 'abspath', 'key'):
             if not isinstance(record[key], str):
-                logging.info('{0} is not a string: {1}, but {2}'.format(key, record[key], record[key].__class__))
+                get_logger().info('{0} is not a string: {1}, but {2}'.format(key, record[key], record[key].__class__))
                 return False
 
         for key in ('size', 'part_size'):
             if not isinstance(record[key], int):
-                logging.info('{0} is not an integer: {1}'.format(key, record[key]))
+                get_logger().info('{0} is not an integer: {1}'.format(key, record[key]))
                 return False
 
         if not isinstance(record['mtime'], int) and not isinstance(record['mtime'], float):
-            logging.info('mtime is not a float or an integer: {0}'.format(record['mtime']))
+            get_logger().info('mtime is not a float or an integer: {0}'.format(record['mtime']))
             return False
 
         if not isinstance(record['parts'], list):
-            logging.info('parts is not a list: {0}'.format(record['parts'].__class__.__name__))
+            get_logger().info('parts is not a list: {0}'.format(record['parts'].__class__.__name__))
             return False
     except KeyError as e:
-        logging.info('Key not found: {0}'.format(e.args))
+        get_logger().info('Key not found: {0}'.format(e.args))
         return False
 
     return True